+++ src/note/asset_base.rs
